+++ libraries/botframework-streaming/src/adapters/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module botframework-streaming
+ */
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from './streamingAdapter';
+export * from './streamingHttpClient';
+export * from './tokenResolver';
